requirements.txt
nltk==3.4
openpyxl==2.5.12
bs4==0.0.1
pyenchant==2.0.0
wikipedia==1.4.0
stanford-corenlp==3.9.2
pandas==0.23.4
Flask==1.0.2
APScheduler==3.5.3
redis==3.0.1